Add 'insights' command to query the NR Insights API
New Relic Insights can be queried by submitting NRQL queries
to an API endpoint. This adds a new hubot command for doing that.

This first attempt supports regular queries that return lists
of events, along with faceted queries that group the results
by an attribute.

Other types of queries, including multi-facet queries, are not
yet supported.