Fix 'cfgov-deployed' queries to not use 'SINCE this quarter'
This corrects a misunderstanding of how 'SINCE this quarter' works in
NRQL. I wrongly assumed it would be similar to 'SINCE 90 days ago',
but it's really since the beginning of the current quarter, which
makes sense. But that's not what we want. We just want to go back
sufficiently far enough, which is what this change gets us.