Update 'cfgov-deployed', use APM instead of Insights
While it was a nicer solution to use NRQL and Insights to get this
information, we can't be sure that our custom Insights events will
stick around long enough (at least not at the present time), and
the 'cfgov-deployed' command's usefulness would be compromised as
a result: deploys might stop showing up as custom events got truncated
in Insights over time.

Now we are updating the command to use the somewhat clunkier APM-based
APIs. There should be no noticeable difference from the user perspective
-- except for one improvement. Now that the deploy target ordering is
explicit in the config file, we can order the results exactly as we
want.